Update RetireApplication.ps1
Nouvelles variables
Ajout de tests pour éviter de copier et supprimer les applications si déjà retired

+++ RetireApplication.ps1
@@ -10,7 +10,7 @@
 #
 # Updates:
 #        0.1 - Raphael Perez - 16/01/2014 - Initial Script
-#        1.0 - io4Master- 27-02-2019 - Modifications
+#        1.0 - io4Master- 22-03-2019 - Modifications
 #
 # Usage:
 #	 Option 1: powershell.exe -ExecutionPolicy Bypass .\RetireApplication.ps1 [Parameters]
@@ -31,10 +31,14 @@ $SiteCode = $SiteNamespace.SubString($SiteNamespace.Indexof("site_") +5)
 $modelname = $args[2]
 $Scope = "Retired" #Path in Sccm Application Where you want to put retired application
 
-$NewRootPath = "\\server\packages$\__Retired"
-$localcachelocation = "C:\ProgramData\Client\Sources\_Uninstall\"
+$CMserver = ""
+$ClientName = ""
+$SRCServer = ""
+
+$NewRootPath = "\\$CMserver\packages$\__Retired"
+$localcachelocation = "C:\ProgramData\$ClientName\Sources\_Uninstall"
 $localPKGExe = "Uninstall.exe"
-$Dummyfolder = "\\serveur\packages$\_UninstallOnly"
+$Dummyfolder = "\\$SRCServer\packages$\_UninstallOnly"
 
 
 try
@@ -72,12 +76,12 @@ try
     $ApplicationXML = [Microsoft.ConfigurationManagement.ApplicationManagement.Serialization.SccmSerializer]::DeserializeFromString($App.SDMPackageXML, $true)
 
     ##Retiring App
-	Write-host "Retiring application"
+	Write-host "Retiring application in SCCM Console"
 	$App.SetIsExpired($true) | out-null
 	Write-host " "
     
     ##Move Application to Retired folder
-	Write-host "Moving Application to Folder $Scope"
+	Write-host "Moving Application to Folder $Scope in SCCM Console"
 	$AppObj = Get-CMApplication -Name $NewName 
     $AppObj | Move-CMObject -FolderPath $SiteCode":\Application\"$Scope
             
@@ -104,24 +108,28 @@ try
         for ($i=0; $i -lt $DeplName.Count ; $i++)
         {
             $Loc=""
-            write-host $AppLocation[$i]
             $Loc=$AppLocation[$i].Substring(0,$AppLocation[$i].Length-1)
             $newPath = "$($NewRootPath)$($Loc.Substring($Loc.LastIndexOf("\")+1))"
-                   
-			Write-host "Creating retired path $newPath"
-			if (!(Test-Path $newPath)) { [system.io.directory]::CreateDirectory($newPath) | out-null }
-			New-PSDrive -Name source -PSProvider FileSystem -Root $AppLocation[$i] | Out-Null
-			New-PSDrive -Name target -PSProvider FileSystem -Root $newPath | Out-Null
-			Write-host "Copying files $($AppLocation[$i]) to $newPath"
-			Copy-Item -Path source:\*.* -Destination target: -recurse
-			Remove-PSDrive source
-			Remove-PSDrive target
+            write-host "Application sources Location $AppLocation[$i]"
+            if ((test-path $newPath) -eq (test-path $($AppLocation[$i]))) { Write-host "Same Paths - Already Retired" | out-null } else {
+                Write-host "Creating retired path $newPath"                  
+			   
+			    if (!(Test-Path $newPath)) { [system.io.directory]::CreateDirectory($newPath) | out-null }
+			    New-PSDrive -Name source -PSProvider FileSystem -Root $AppLocation[$i] | Out-Null
+			    New-PSDrive -Name target -PSProvider FileSystem -Root $newPath | Out-Null
+			    Write-host "Copying files $($AppLocation[$i]) to $newPath"
+                Copy-Item -Path source:\*.* -Destination target: -recurse
+                
+			    Remove-PSDrive source
+			    Remove-PSDrive target
+                }
+            
 
 
 			##Change location in XML (Local Cache)
             #$newuninstallpath = join-path($localcachelocation,$DeplName)
             
-            Write-host "Changing folder source to $Dummyfolder"
+            Write-host "Changing application source folder to $Dummyfolder"
             $ApplicationXML.DeploymentTypes.Installer[$i].Contents[0].Location = $Dummyfolder+"\"
 
             $ApplicationXML.DeploymentTypes.Installer[$i].UninstallCommandLine = $localcachelocation+$DeplName[$i].'#text'+"\"+$localPKGExe
@@ -129,18 +137,21 @@ try
             $ApplicationXML.DeploymentTypes.Installer[$i].UninstallSetting="NoneRequired"
 
 
-			Write-host "Deleting original source folder $($AppLocation[$i])"
 			#delete source directory
-			New-PSDrive -Name source -PSProvider FileSystem -Root ($AppLocation[$i].Substring(0,$AppLocation[$i].LastIndexOf("\"))) | Out-Null
-				    
-            Get-ChildItem -Path source:\"$($AppLocation[$i].Substring($AppLocation[$i].LastIndexOf("\")+1))" -Recurse | Remove-Item -force -Recurse
-			remove-item -Path source:\"$($AppLocation[$i].Substring($AppLocation[$i].LastIndexOf("\")+1))" -Force
-			Remove-PSDrive source
+			
+            if ((test-path $newPath) -eq (test-path $($AppLocation[$i]))) { Write-host "Same Paths - Already Retired" | out-null } else { 
+                Write-host "Deleting original source folder $($AppLocation[$i])"
+                New-PSDrive -Name source -PSProvider FileSystem -Root ($AppLocation[$i].Substring(0,$AppLocation[$i].LastIndexOf("\"))) | Out-Null
+                Get-ChildItem -Path source:\"$($AppLocation[$i].Substring($AppLocation[$i].LastIndexOf("\")+1))" -Recurse | Remove-Item -force -Recurse
+			    remove-item -Path source:\"$($AppLocation[$i].Substring($AppLocation[$i].LastIndexOf("\")+1))" -Force
+                Remove-PSDrive source
+                }
+			
         }
-    
+    write-host "Update Application Properties"
     $UpdatedXML = [Microsoft.ConfigurationManagement.ApplicationManagement.Serialization.SccmSerializer]::SerializeToString($ApplicationXML, $true)
 	$App.SDMPackageXML = $UpdatedXML
-	$App.Put()
+	#$App.Put()
 	}
 
 
@@ -171,7 +182,7 @@ try
             $DPName = (Get-CMDistributionPointInfo | ? { $_.NALPath -eq $DPinfo.ServerNALPath}).Name
             Write-host "Removing ""$NewName"" package content from distribution point:"$DPName
 
-            Remove-CMContentDistribution -ApplicationName "$NewName" -DistributionPointName "$DPName" -Force 
+            #Remove-CMContentDistribution -ApplicationName "$NewName" -DistributionPointName "$DPName" -Force 
         }
     }
